atomic pattern
