+++ src/components/atoms/Form.ts
@@ -1,6 +1,6 @@
 import { styled } from 'styled-components';
 
-export const StForm = styled.form`
+export const MainForm = styled.form`
   display: flex;
   align-items: center;
   justify-content: center;
