+++ src/components/molecules/Footer.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { FooterContainer } from '@components/atoms/Container';
+import NavigationMenu from '@components/molecules/NavigationMenu';
+
+const Footer = () => {
+  return (
+    <FooterContainer>
+      <NavigationMenu />
+    </FooterContainer>
+  );
+};
+
+export default Footer;
