+++ src/components/molecules/LodingIndicator.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { ClipLoader } from 'react-spinners';
+import { LoadingIndicatorContainer } from '@components/atoms/Container';
+
+const LoadingIndicator = () => {
+  return (
+    <LoadingIndicatorContainer>
+      <ClipLoader size={80} color={'var(--color-main)'} loading={true} />
+    </LoadingIndicatorContainer>
+  );
+};
+
+export default LoadingIndicator;
