--- src/components/types.tsx
@@ -1,3 +1,0 @@
-export type LoadingProps = {
-  setLoading: React.Dispatch<React.SetStateAction<boolean>>;
-};
