+++ src/core/Router.tsx
@@ -4,8 +4,8 @@ import Start from '@pages/Start';
 import Login from '@pages/Login';
 import Signup from '@pages/Signup';
 import Mypage from '@pages/Mypage';
-import Layout from '@layout/Layout';
-import LoginLayout from '@layout/LoginLayout';
+import Layout from '@components/molecules/MainLayout';
+import LoginLayout from '@components/molecules/StartLayout';
 import Main from '@pages/Main';
 import Search from '@pages/Search';
 import Diary from '@pages/Diary';
@@ -13,7 +13,7 @@ import DailyLog from '@pages/DailyLog';
 import DailyLogAdd from '@pages/DailyLogAdd';
 
 import { Toaster } from 'react-hot-toast';
-import LoadingIndicator from '@components/LodingIndicator';
+import LoadingIndicator from '@components/molecules/LodingIndicator';
 
 const Router = () => {
   const [loading, setLoading] = useState(false);
